tweaked 'deleteConversation' controller function deletion logic to prevent deleting the conversation entirely if the user who sent the DM request clears it from their list of convos

+++ api/controllers/message-related/conversation-related.ts
@@ -152,16 +152,18 @@ const deleteConversation = async (
 						dmRequests: conversationID
 					}
 				});
-				console.log("ran");
-			}
-
-			await User.findByIdAndUpdate(currUID, {
-				$pull: {
-					conversations: conversationID
+			} else {
+				if (conversation.isDMRequest) {
+					// if the conversation is a DM request, then remove it from the requestedBy user's dmRequests
+					await User.findByIdAndUpdate(currUID, {
+						$pull: {
+							conversations: conversationID
+						}
+					});
+				} else {
+					await Conversation.findByIdAndDelete(conversationID);
 				}
-			});
-
-			await Conversation.findByIdAndDelete(conversationID);
+			}
 		} else {
 			await User.findByIdAndUpdate(currUID, {
 				$pull: {
